refactor(runtime): extract provider call into agent/turn — #7415
Move the LLM-request announcement (Thinking status, observer LlmRequest,
llm_request log, fire_llm_input hook), budget enforcement, and the
provider call (streaming via consume_provider_streaming_response with
non-streaming fallback; non-streaming with optional step timeout and
cancel select) into turn/provider_call.rs as announce_llm_request(..),
enforce_tool_loop_budget(), and call_provider(..) ->
Result<ProviderCallOutcome>.

The cancel asymmetry is preserved verbatim and now encoded in the type:
the non-streaming cancel paths and the step-timeout bails return
call_provider's OUTER Err (the loop propagates with ?, skipping
observer-failure recording and context-overflow recovery, as today),
while the streaming-fallback cancel remains an Err as the chat_result
VALUE (ProviderCallOutcome.chat_result), flowing through the loop's
match Err arm exactly as before.

RUN_SHEET deviations recorded in the same change: announce_llm_request
gains the active provider param (the llm_request attribution_span must
reference the active — possibly vision — provider, which is not in
TurnCtx) and returns llm_started_at (upstream takes the Instant between
the log line and the hook; returning it preserves exact timing
semantics).

The request_tools / should_consume_provider_stream derivation stays
inline in the loop. ChatRequest and the consume_provider_streaming_response
re-export are now test-only in loop_.rs; BudgetCheck import dropped
(all lib users moved).
